Bump scala version to 2.13 (#2336)
This PR bumps Scala language from 2.12 to 2.13 according to the [migration guide](https://docs.scala-lang.org/overviews/core/collections-migration-213.html) with some dependency upgrades.

The following dependency requires Java 11+ for their latest version, so we don't update them since we still want to support Java 8:
1. Lucene
2. JGraphT
3. Logback
4. Mariadb4j
5. Jooq
6. Slf4j

+++ core/amber/src/main/scala/edu/uci/ics/texera/web/auth/JwtAuth.scala
@@ -43,7 +43,7 @@ object JwtAuth {
     claims.setClaim("googleId", user.getGoogleId)
     claims.setClaim("email", user.getEmail)
     claims.setClaim("role", user.getRole)
-    claims.setExpirationTimeMinutesInTheFuture(dayToMin(expireInDays))
+    claims.setExpirationTimeMinutesInTheFuture(dayToMin(expireInDays).toFloat)
     claims
   }
